some naming nits, small restructure of poly-analysis folder and other nits
