--- feature-utils/poly-analysis/src/analysis/generic-analyses/generic-analyses.js
@@ -1,9 +1,0 @@
-import DataStructureBubblesAnalysis from "./data-structure-bubbles-analysis";
-import DataImportingStatusAnalysis from "./importing-status-analysis";
-
-const genericAnalyses = [
-    DataStructureBubblesAnalysis,
-    DataImportingStatusAnalysis,
-];
-
-export default genericAnalyses;
