--- feature-utils/poly-analysis/src/analysis/report/report-analysis.js
@@ -1,14 +1,0 @@
-import RootAnalysis from "../root-analysis.js";
-
-export default class ReportAnalysis extends RootAnalysis {
-    get isForDataReport() {
-        return true;
-    }
-
-    get jsonReport() {
-        return {
-            id: this.id,
-            data: this.reportData,
-        };
-    }
-}
