+++ feature-utils/poly-analysis/src/generic-analyses.js
@@ -1,0 +1,9 @@
+import DataStructureBubblesAnalysis from "./analysis/data-structure-bubbles-analysis";
+import DataImportingStatusReport from "./report/data-importing-status-report";
+
+const genericAnalyses = [
+    DataStructureBubblesAnalysis,
+    DataImportingStatusReport,
+];
+
+export default genericAnalyses;
