+++ feature-utils/poly-analysis/src/report/data-importing-status-report.js
@@ -1,6 +1,6 @@
-import ReportAnalysis from "../report/report-analysis";
+import ReportAnalysis from "./report-analysis";
 
-export default class DataImportingStatusAnalysis extends ReportAnalysis {
+export default class DataImportingStatusReport extends ReportAnalysis {
     async analyze({ dataAccount }) {
         dataAccount.reports.importersData = dataAccount.importingResults.map(
             (importerResult) => importerResult.reportJsonData
