+++ features/google/src/views/ministories/reports.js
@@ -1,0 +1,5 @@
+import DataImportingStatusReport from "./dataImportingStatus.jsx";
+
+export const reports = [DataImportingStatusReport];
+
+export default reports;
