+++ features/google/src/views/overview/overview.jsx
@@ -26,7 +26,7 @@ const Overview = () => {
         analyzeFile({
             zipData: files[0],
             dataAccount: account,
-            subAnalyses: [],
+            specificAnalyses: [],
         });
     }, [account]);
